Adding new git-blame command, moving CatFactsSubscribers json to blob storage, and making sure to sync commands between all active guilds

+++ ShaosilBot/DependencyInjection/DiscordSocketClientProvider.cs
@@ -6,7 +6,6 @@ using ShaosilBot.SlashCommands;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Net.Http;
 using System.Threading.Tasks;
 
 namespace ShaosilBot.DependencyInjection
@@ -15,13 +14,9 @@ namespace ShaosilBot.DependencyInjection
     {
         private readonly static DiscordSocketConfig _config = new DiscordSocketConfig { GatewayIntents = GatewayIntents.GuildMembers | GatewayIntents.DirectMessages };
         private readonly static DiscordSocketClient _client = new DiscordSocketClient(_config);
-        private static HttpClient _httpClient;
 
         public static DiscordSocketClient GetSocketClient(IServiceProvider provider)
         {
-            if (_httpClient == null)
-                _httpClient = provider.GetRequiredService<IHttpClientFactory>().CreateClient();
-
             if (_client.LoginState == LoginState.LoggedOut)
             {
                 var logger = provider.GetService<ILogger<DiscordSocketClient>>();
@@ -30,9 +25,7 @@ namespace ShaosilBot.DependencyInjection
                 _client.Ready += async () =>
                 {
                     await KeepAlive();
-
-                    // Uncomment to refresh commands
-                    //await SyncCommands();
+                    await SyncCommands();
                 };
                 _client.MessageReceived += MessageHandler;
 
@@ -53,36 +46,40 @@ namespace ShaosilBot.DependencyInjection
             if (socketMessage.Channel.GetChannelType() == ChannelType.DM && !socketMessage.Author.IsBot && socketMessage.Content.ToUpper().Contains("STOP"))
             {
                 // Update times unsubscribed lol
-                var currentSubscribers = await CatFactsCommand.GetSubscribersAsync(_httpClient);
+                var currentSubscribers = await CatFactsCommand.GetSubscribersAsync();
                 var matchingSubscriber = currentSubscribers.FirstOrDefault(s => s.IDNum == socketMessage.Author.Id);
                 if (matchingSubscriber != null)
                 {
                     matchingSubscriber.TimesUnsubscribed++;
-                    await CatFactsCommand.UpdateSubscribersAsync(currentSubscribers, _httpClient);
+                    await CatFactsCommand.UpdateSubscribersAsync(currentSubscribers);
                 }
                 string extraMessage = matchingSubscriber?.TimesUnsubscribed > 1 ? $" (Wowza! You have subscribed {matchingSubscriber.TimesUnsubscribed} times!) " : string.Empty;
-                //await foreach (var message in socketMessage.Channel.GetMessagesAsync().Flatten())
-                //{
-                //    if (message.Author.IsBot)
-                //        await message.DeleteAsync();
-                //}
                 await socketMessage.Author.SendMessageAsync($"Thanks for subscribing to Cat Facts Digest (CFD){extraMessage}! Be prepared to boost that feline knowledge every hour, on the hour, between the hours of 10:00 AM and 10:00 PM EST! *Meow!*");
             }
         }
 
         private async static Task SyncCommands()
         {
-            var guild = _client.GetGuild(628019972316069890);
-            await guild.DeleteApplicationCommandsAsync();
-            await guild.CreateApplicationCommandAsync(new SlashCommandBuilder { Name = "test-command", Description = "Getting closer to world domination", DefaultMemberPermissions = GuildPermission.Administrator }.Build());
-            await guild.CreateApplicationCommandAsync(new SlashCommandBuilder { Name = "wow", Description = "Wow." }.Build());
-            await guild.CreateApplicationCommandAsync(new SlashCommandBuilder { Name = "cat-fact", Description = "Thank you for subscribing to cat facts! Text STOP to unsubscribe." }.Build());
-            await guild.CreateApplicationCommandAsync(new SlashCommandBuilder
+            var guilds = _client.Guilds;
+
+            foreach (var guild in guilds)
             {
-                Name = "xkcd",
-                Description = "Get a random XKCD comic, or optionally a specific one!",
-                Options = new List<SlashCommandOptionBuilder> { new SlashCommandOptionBuilder { Name = "comic-num", Type = ApplicationCommandOptionType.Integer, MinValue = 0, Description = "The number of the comic to pull. 0 for current. Omit for random." } }
-            }.Build());
+                // For now, just create all commands if the guild has none
+                var existingCommands = await guild.GetApplicationCommandsAsync();
+                if (existingCommands.Count > 0) continue;
+
+                await guild.DeleteApplicationCommandsAsync();
+                await guild.CreateApplicationCommandAsync(new SlashCommandBuilder { Name = "test-command", Description = "Getting closer to world domination", DefaultMemberPermissions = GuildPermission.Administrator }.Build());
+                await guild.CreateApplicationCommandAsync(new SlashCommandBuilder { Name = "wow", Description = "Wow." }.Build());
+                await guild.CreateApplicationCommandAsync(new SlashCommandBuilder { Name = "cat-fact", Description = "Thank you for subscribing to cat facts! Text STOP to unsubscribe." }.Build());
+                await guild.CreateApplicationCommandAsync(new SlashCommandBuilder
+                {
+                    Name = "xkcd",
+                    Description = "Get a random XKCD comic, or optionally a specific one!",
+                    Options = new List<SlashCommandOptionBuilder> { new SlashCommandOptionBuilder { Name = "comic-num", Type = ApplicationCommandOptionType.Integer, MinValue = 0, Description = "The number of the comic to pull. 0 for current. Omit for random." } }
+                }.Build());
+                await guild.CreateApplicationCommandAsync(new SlashCommandBuilder { Name = "git-blame", Description = "Blame a random one of (Shaosil, Syrelash, mbmminer, or Skom)" }.Build());
+            }
         }
     }
 }
